Consolidate BSD shadow support into single module
Since OpenBSD works the same way as NetBSD, instead of writing a third
module this commit consolidates BSD shadow operations into a single
module.
